feat(heater-shaker): Heater support (#168)
- pid uses doubles instead of floats
- Moves heater hal handles to internal struct typedef
- Adds power setting to heater control policy
- Integrates PID control. The heater now has a PID controller that it uses to generate power settings through the policy. This runs when a new temperature reading is received, and only if the temperature setting does not incur an error. This means the temperature read frequency will drive the controller regulation frequency.
- Adds M301, set-pid-constant. You can use M301 to set the PID constants for the heater. They won't be
saved to flash, but will last until the next set (or until restart).
- Adds M104.D, set-heater-power, to set a specific relative heater power in (0, 1) for min->max.

+++ stm32-modules/heater-shaker/src/pid.cpp
@@ -3,12 +3,12 @@
 #include <algorithm>
 #include <limits>
 
-PID::PID(float kp, float ki, float kd)
-    : PID(kp, ki, kd, std::numeric_limits<float>::infinity(),
-          -std::numeric_limits<float>::infinity()) {}
+PID::PID(double kp, double ki, double kd)
+    : PID(kp, ki, kd, std::numeric_limits<double>::infinity(),
+          -std::numeric_limits<double>::infinity()) {}
 
-PID::PID(float kp, float ki, float kd, float windup_limit_high,
-         float windup_limit_low)
+PID::PID(double kp, double ki, double kd, double windup_limit_high,
+         double windup_limit_low)
     : _kp(kp),
       _ki(ki),
       _kd(kd),
@@ -17,24 +17,24 @@ PID::PID(float kp, float ki, float kd, float windup_limit_high,
       _integrator(0),
       _last_error(0) {}
 
-auto PID::kp() const -> float { return _kp; }
+auto PID::kp() const -> double { return _kp; }
 
-auto PID::ki() const -> float { return _ki; }
+auto PID::ki() const -> double { return _ki; }
 
-auto PID::kd() const -> float { return _kd; }
+auto PID::kd() const -> double { return _kd; }
 
-auto PID::windup_limit_high() const -> float { return _windup_limit_high; }
+auto PID::windup_limit_high() const -> double { return _windup_limit_high; }
 
-auto PID::windup_limit_low() const -> float { return _windup_limit_low; }
+auto PID::windup_limit_low() const -> double { return _windup_limit_low; }
 
-auto PID::integrator() const -> float { return _integrator; }
+auto PID::integrator() const -> double { return _integrator; }
 
-auto PID::last_error() const -> float { return _last_error; }
+auto PID::last_error() const -> double { return _last_error; }
 
-auto PID::compute(float error) -> float {
+auto PID::compute(double error) -> double {
     _integrator =
         std::clamp(error + _integrator, _windup_limit_low, _windup_limit_high);
-    const float errdiff = error - _last_error;
+    const double errdiff = error - _last_error;
     _last_error = error;
     return (_kp * error) + (_kd * errdiff) + (_ki * _integrator);
 }
